T3C-74 make create report form simpler
