+++ next-client/src/components/create/components/Modals.tsx
@@ -1,0 +1,77 @@
+import React from "react";
+import {
+  AlertDialog,
+  AlertDialogContent,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogCancel,
+  AlertDialogAction,
+} from "@/components/elements";
+import { Button } from "@/components/elements";
+import { signInWithGoogle } from "@/lib/firebase/auth";
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogHeader,
+  DialogTitle,
+} from "@/components/elements";
+
+export function PoorlyFormattedModal({
+  isOpen,
+  cancelFunc,
+  proceedFunc,
+}: {
+  isOpen: boolean;
+  cancelFunc: () => void;
+  proceedFunc: () => void;
+}) {
+  return (
+    <AlertDialog open={isOpen}>
+      <AlertDialogContent className="w-[329px] gap-6">
+        <AlertDialogHeader>
+          <AlertDialogTitle>
+            We detected a poorly formatted csv. Do you want to proceed?
+          </AlertDialogTitle>
+          <AlertDialogDescription>
+            Please make sure the CSV contains at least two columns named "id"
+            and "comment"
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel onClick={cancelFunc} asChild>
+            <Button variant={"outline"}>Cancel</Button>
+          </AlertDialogCancel>
+          <AlertDialogAction onClick={proceedFunc} className="bg-destructive">
+            Proceed
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  );
+}
+
+export const SigninModal = ({ isOpen }: { isOpen: boolean }) => {
+  const handleSignIn = async () => {
+    try {
+      await signInWithGoogle();
+    } catch (error) {
+      console.error("Sign in failed:", error);
+    }
+  };
+
+  return (
+    <Dialog open={isOpen}>
+      <DialogContent className="gap-10">
+        <DialogHeader>
+          <DialogTitle>Login to create a report</DialogTitle>
+        </DialogHeader>
+        <DialogDescription>
+          <Button onClick={handleSignIn}>Login</Button>
+        </DialogDescription>
+      </DialogContent>
+    </Dialog>
+  );
+};
